Fix windows/mingw build.
fbbd41 (Pull request #16) added g_autofree, which is not supported on the old version of GTK in the pidgin-windev.

Fix a size_t printf format (on Windows, size_t is not necessarily the same as unsigned long)

Fix a duplicate symbol at link by declaring HFUJabberStreamDataTable extern in the header, and only defining it in jabber_http_file_upload.c.